backend/requirements.txt
Flask
flask-cors
scikit-learn
pandas
nltk
joblib
python-docx
openpyxl
requests
